requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.8.0
blinker==1.4
cffi==1.14.2
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pycparser==2.20
six==1.15.0
soupsieve==1.9.3
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
